config/requirements.txt
asgiref==3.8.1
clamd==1.0.2
Django==5.2.3
django-safe-filefield==1.0.0
django-taggit==6.1.0
djangorestframework==3.16.0
drf-nested-routers==0.94.2
psycopg2-binary==2.9.10
python-magic==0.4.27
python-magic-bin==0.4.14
sqlparse==0.5.3
tzdata==2025.2
